requirements.txt
Flask==3.1.0
groq==0.30.0
PyPDF2==3.0.1
python-docx==1.2.0
sentence-transformers==5.0.0
faiss-cpu==1.11.0.post1
numpy==2.2.1
pandas==2.2.3

